- matrix mult int overflow

+++ benchmarks_new/matrix_mult/matrix_mult_main.h
@@ -20,17 +20,17 @@ double CallKernel(int core_type)
 	static array_type a;
 	static array_type b;
 	static array_type c;
-    int bw_for_stat[6] = {(int)LENGTH * (int)LENGTH* (int)LENGTH * 2  + 2 * (int)LENGTH * (int)LENGTH,
-                              (int)LENGTH * (int)LENGTH* (int)LENGTH * 3  + 1 * (int)LENGTH * (int)LENGTH,
-                              (int)LENGTH * (int)LENGTH* (int)LENGTH * 2  + 2 * (int)LENGTH * (int)LENGTH,
-                              (int)LENGTH * (int)LENGTH* (int)LENGTH * 3  + 1 * (int)LENGTH * (int)LENGTH,
-                              (int)LENGTH * (int)LENGTH* (int)LENGTH * 3  + 1 * (int)LENGTH * (int)LENGTH,
-                              (int)LENGTH * (int)LENGTH* (int)LENGTH * 3  + 1 * (int)LENGTH * (int)LENGTH};
+    int bw_for_stat[6] = {(double)LENGTH * (double)LENGTH* (double)LENGTH * 2  + 2 * (double)LENGTH * (double)LENGTH,
+                              (double)LENGTH * (double)LENGTH* (double)LENGTH * 3  + 1 * (double)LENGTH * (double)LENGTH,
+                              (double)LENGTH * (double)LENGTH* (double)LENGTH * 2  + 2 * (double)LENGTH * (double)LENGTH,
+                              (double)LENGTH * (double)LENGTH* (double)LENGTH * 3  + 1 * (double)LENGTH * (double)LENGTH,
+                              (double)LENGTH * (double)LENGTH* (double)LENGTH * 3  + 1 * (double)LENGTH * (double)LENGTH,
+                              (double)LENGTH * (double)LENGTH* (double)LENGTH * 3  + 1 * (double)LENGTH * (double)LENGTH};
 
 
 	double time = -1;
     size_t bytes_requested = sizeof(double) * bw_for_stat[core_type];
-    size_t flops_requested = (int)LENGTH * (int)LENGTH* (int)LENGTH * 2;
+    size_t flops_requested = (double)LENGTH * (double)LENGTH* (double)LENGTH * 2;
     auto counter = PerformanceCounter();
 
 	for(int i = 0; i < LOC_REPEAT; i++)
